first draft of GuestExcel class done
